+++ src/main/java/com/github/zk1931/pulsed/HttpWatch.java
@@ -17,6 +17,8 @@
 
 package com.github.zk1931.pulsed;
 
+import com.github.zk1931.pulsed.tree.Node;
+import com.github.zk1931.pulsed.tree.Watch;
 import java.io.IOException;
 import javax.servlet.AsyncContext;
 import javax.servlet.http.HttpServletResponse;
