+++ src/main/java/com/github/zk1931/pulsed/Pulsed.java
@@ -17,6 +17,17 @@
 
 package com.github.zk1931.pulsed;
 
+import com.github.zk1931.jzab.PendingRequests;
+import com.github.zk1931.jzab.StateMachine;
+import com.github.zk1931.jzab.Zab;
+import com.github.zk1931.jzab.ZabConfig;
+import com.github.zk1931.jzab.ZabException;
+import com.github.zk1931.jzab.ZabException.NotBroadcastingPhase;
+import com.github.zk1931.jzab.ZabException.TooManyPendingRequests;
+import com.github.zk1931.jzab.Zxid;
+import com.github.zk1931.pulsed.tree.DataTree;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import java.io.InputStream;
 import java.io.IOException;
 import java.io.OutputStream;
@@ -28,16 +39,6 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 import java.util.Set;
 import javax.servlet.AsyncContext;
-import com.github.zk1931.jzab.PendingRequests;
-import com.github.zk1931.jzab.StateMachine;
-import com.github.zk1931.jzab.Zab;
-import com.github.zk1931.jzab.ZabConfig;
-import com.github.zk1931.jzab.ZabException;
-import com.github.zk1931.jzab.ZabException.NotBroadcastingPhase;
-import com.github.zk1931.jzab.ZabException.TooManyPendingRequests;
-import com.github.zk1931.jzab.Zxid;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * State machine.
