+++ src/main/java/com/github/zk1931/pulsed/PulsedSessionsHandler.java
@@ -18,6 +18,10 @@
 package com.github.zk1931.pulsed;
 
 import com.github.zk1931.jzab.ZabException;
+import com.github.zk1931.pulsed.tree.DataTree;
+import com.github.zk1931.pulsed.tree.FileNode;
+import com.github.zk1931.pulsed.tree.Node;
+import com.github.zk1931.pulsed.tree.PathUtils;
 import java.io.IOException;
 import java.nio.charset.Charset;
 import java.util.Arrays;
