+++ src/main/java/com/github/zk1931/pulsed/RemoveCommand.java
@@ -17,6 +17,7 @@
 
 package com.github.zk1931.pulsed;
 
+import com.github.zk1931.pulsed.tree.Node;
 import javax.servlet.AsyncContext;
 import javax.servlet.http.HttpServletResponse;
