+++ src/main/java/com/github/zk1931/pulsed/TreeHandler.java
@@ -18,6 +18,9 @@
 package com.github.zk1931.pulsed;
 
 import com.github.zk1931.jzab.ZabException;
+import com.github.zk1931.pulsed.tree.DataTree;
+import com.github.zk1931.pulsed.tree.PathUtils;
+import com.github.zk1931.pulsed.tree.Node;
 import java.io.IOException;
 import javax.servlet.AsyncContext;
 import javax.servlet.AsyncEvent;
